+++ Assets/Scripts/AutoSave.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+using UnityEngine.SceneManagement;
+
+public class AutoSave : MonoBehaviour
+{
+
+    // Update is called once per frame
+    void Update()
+    {
+        PlayerPrefs.SetInt("SavedScene", SceneManager.GetActiveScene().buildIndex);
+    }
+}
