+++ Assets/Scripts/AutoSave.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 87a60fe9379a93344817cc953fcdd39f
+guid: 41c68b81fa679ba418176e796f0ce154
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
