+++ Assets/Scripts/InGameMenu.cs
@@ -2,6 +2,8 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+using UnityEngine.SceneManagement;
+
 public class InGameMenu : MonoBehaviour{
     
     public static bool GameIsPaused = false;
@@ -28,5 +30,16 @@ public class InGameMenu : MonoBehaviour{
         pauseMenu.SetActive(true);
         Time.timeScale = 0f;
         GameIsPaused = true;
+    }
+
+    public void StartNextLevel(){
+        SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex + 1);
+        PlayerPrefs.SetInt("SavedScene", SceneManager.GetActiveScene().buildIndex);
+    }
+
+    public void QuitToMainMenu(){
+        Resume();
+        PlayerPrefs.SetInt("SavedScene", SceneManager.GetActiveScene().buildIndex);
+        SceneManager.LoadScene("Menu");
     }
 }
