+++ Assets/Scripts/Save.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: fbfd5735a256aca458d2e94201b7f257
+guid: 2179c44029417464d9dbc1e7b5055789
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
