+++ Assets/Scripts/Save/Player.cs
@@ -2,10 +2,16 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+using UnityEngine.SceneManagement;
+
 public class Player : MonoBehaviour{
     
-    public int level = 2;
-    public int health = 12;
+    public string level;
+
+    void Update(){
+      level = SceneManager.GetActiveScene().name;
+      SavePlayer();
+    }
 
     public void SavePlayer(){
         SaveSystem.SavePlayer(this);
@@ -15,12 +21,6 @@ public class Player : MonoBehaviour{
         PlayerData data = SaveSystem.LoadPlayer(this);
 
         level = data.level;
-        health = data.health;
-
-        Vector2 position;
-        position.x = data.position[0];
-        position.y = data.position[1];
-        transform.position = position;
     }
 
       public void RestartLostGame(){
