+++ Assets/Scripts/Save/PlayerData.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[System.Serializable]
+public class PlayerData{
+    
+    public string level;
+
+    public PlayerData(Player player){
+        level = player.level;
+    }
+}
